IO componet is completely refactored, most of unit tests are implemented

+++ src/main/java/edu/sanekas/cashmachine/api/CashManipulator.java
@@ -6,8 +6,8 @@ import edu.sanekas.wrapper.api.InputWrapper;
 import java.util.Map;
 
 public interface CashManipulator {
-    int putCash(InputWrapper inputWrapper);
-    Map<Nominal, Integer> getCash(InputWrapper inputWrapper);
-    Map<Nominal, Integer> dump();
-    int state();
+    long putCash(InputWrapper inputWrapper);
+    Map<Nominal, Long> getCash(InputWrapper inputWrapper);
+    Map<Nominal, Long> dump();
+    long state();
 }
