+++ src/main/java/edu/sanekas/command/impl/commands/GetCommand.java
@@ -20,10 +20,10 @@ public class GetCommand implements Command {
 
     @Override
     public OutputWrapper execute(InputWrapper inputWrapper) {
-        Map<Nominal, Integer> cash = cashManipulator.getCash(inputWrapper);
-        OutputWrapper<Map<Nominal, Integer>> outputWrapper = wrapperFactory.createOutputWrapper();
-        int receivedCash = 0;
-        for (Map.Entry<Nominal, Integer> cashPair : cash.entrySet()) {
+        Map<Nominal, Long> cash = cashManipulator.getCash(inputWrapper);
+        OutputWrapper<Map<Nominal, Long>> outputWrapper = wrapperFactory.createOutputWrapper();
+        long receivedCash = 0;
+        for (Map.Entry<Nominal, Long> cashPair : cash.entrySet()) {
             receivedCash += cashPair.getValue() * cashPair.getKey().getNominal();
         }
         cash.put(Nominal.ANY, receivedCash);
@@ -32,16 +32,16 @@ public class GetCommand implements Command {
         return outputWrapper;
     }
 
-    private void prepareOutput(OutputWrapper<Map<Nominal, Integer>> outputWrapper, int requestedCash) {
+    private void prepareOutput(OutputWrapper<Map<Nominal, Long>> outputWrapper, long requestedCash) {
         StringBuilder result = new StringBuilder();
         outputWrapper.getWrappedEntity().entrySet().stream()
                 .filter(cashPair -> cashPair.getKey() != Nominal.ANY && cashPair.getValue() != 0)
                 .forEach(cashPair -> result.append(cashPair.getKey()).append("=").append(cashPair.getValue()).append(","));
 
-        int totalReceivedCash = outputWrapper.getWrappedEntity().get(Nominal.ANY);
+        long totalReceivedCash = outputWrapper.getWrappedEntity().get(Nominal.ANY);
         result.append(" total ").append(totalReceivedCash);
 
-        int dif = totalReceivedCash - requestedCash;
+        long dif = totalReceivedCash - requestedCash;
 
         if (dif < 0) {
             result.append("\n" + "without ").append(String.valueOf(-dif));
