+++ src/main/java/edu/sanekas/wrapper/api/WrapperFactory.java
@@ -4,7 +4,7 @@ import edu.sanekas.api.Nominal;
 import edu.sanekas.api.Operation;
 
 public interface WrapperFactory {
-    InputWrapper createInputWrapper(Operation operation, Nominal nominal, Integer cash);
+    InputWrapper createInputWrapper(Operation operation, Nominal nominal, Long cash);
     InputWrapper createInputWrapper(Operation operation);
     <T> OutputWrapper<T> createOutputWrapper();
 }
